fix: refactor more interfaces
