+++ SplineMiner/TrackRenderer.cs
@@ -1,0 +1,29 @@
+using Microsoft.Xna.Framework;
+using Microsoft.Xna.Framework.Graphics;
+using System.Collections.Generic;
+using SplineMiner.Core;
+
+namespace SplineMiner
+{
+    public class TrackRenderer : ITrackRenderer
+    {
+        private readonly Texture2D _pointTexture;
+        private readonly List<Vector2> _debugPoints = [];
+        private readonly bool _enableDebugVisualization = true;
+
+        public TrackRenderer(Texture2D pointTexture)
+        {
+            _pointTexture = pointTexture;
+        }
+
+        public void Draw(SpriteBatch spriteBatch, ITrack track)
+        {
+            // Implementation of track drawing
+        }
+
+        public void DrawDebugInfo(SpriteBatch spriteBatch, float distance, Texture2D debugTexture)
+        {
+            // Implementation of debug visualization
+        }
+    }
+}
